requirements.txt
attrs==16.3.0
click==6.7
python-telegram-bot==5.3.0
psutil==5.2.2